Fixing catalogs for upstream
