+++ cfme/services/catalogs/catalog.py
@@ -1,10 +1,12 @@
+import cfme.fixtures.pytest_selenium as sel
 import cfme.web_ui.accordion as accordion
+import cfme.web_ui.flash as flash
 import cfme.web_ui as web_ui
 import cfme.web_ui.menu  # load the menu nav endpoints
 import cfme.web_ui.toolbar as tb
 import ui_navigate as nav
 import functools
-import cfme.fixtures.pytest_selenium as sel
+from utils.update import Updateable
 
 assert cfme.web_ui.menu  # to placate flake8 (otherwise menu import is unused)
 
@@ -33,19 +35,12 @@ item_form = web_ui.Form(
      ('display_checkbox', "//input[@id='display']"),
      ('add_button', "//img[@title='Add']")])
 
-catalog_tree = web_ui.Tree('//div[@id="stcat_tree_box"]//table')
-
-
-def catalog_in_table(catalog):
-    return "//div[@class='objbox']//td[.='%s']" % catalog.name
-
-
-def catalog_in_tree(catalog):
-    return "//div[@id='stcat_tree_div']//td[@class='standartTreeRow']/span[.='%s']" % catalog.name
+catalog_tree = sel.ver_pick({'default': web_ui.Tree('//div[@id="stcat_tree_box"]//table'),
+                             '9.9.9.9': web_ui.Tree('//div[@id="stcat_treebox"]//ul')})
 
 
 def _all_catalogs_add_new(_):
-    sel.click("//div[@id='stcat_tree_div']//td[.='All Catalogs']")
+    catalog_tree.click_path('All Catalogs')
     tb_select('Add a New Catalog')
 
 
@@ -55,10 +50,35 @@ nav.add_branch(
                   {'catalog_new': _all_catalogs_add_new,
                    'catalog': [lambda ctx: catalog_tree.click_path('All Catalogs',
                                                                    ctx['catalog'].name),
-                               {'catalog_edit': nav.partial(tb_select, "Edit this Item")}]}],
-     'catalog_items': [nav.partial(accordion.click, 'Catalog Items'),
-                       {'catalog_item_new': _all_catalogs_add_new,
-                        'catalog_item': [lambda ctx:
-                                         sel.click(catalog_in_tree(ctx['catalog_item'])),
-                                         {'catalog_item_edit':
-                                          nav.partial(tb_select, "Edit this Item")}]}]})
+                               {'catalog_edit': nav.partial(tb_select, "Edit this Item")}]}]})
+
+
+class Catalog(Updateable):
+    """Represents a Catalog"""
+
+    def __init__(self, name=None, description=None, items=None):
+        self.name = name
+        self.description = description
+        self.items = items
+
+    def create(self):
+        sel.force_navigate('catalog_new')
+        web_ui.fill(form, {'name_text': self.name,
+                           'description_text': self.description,
+                           'button_multiselect': self.items},
+                    action=form.add_button)
+        flash.assert_success_message('ServiceTemplateCatalog "{}" was saved'.format(self.name))
+
+    def update(self, updates):
+        sel.force_navigate('catalog_edit', context={'catalog': self})
+        web_ui.fill(form, {'name_text': updates.get('name', None),
+                           'description_text': updates.get('description', None),
+                           'button_multiselect': updates.get('items', None)},
+                    action=form.save_button)
+        flash.assert_success_message('ServiceTemplateCatalog "{}" was saved'.format(self.name))
+
+    def delete(self):
+        sel.force_navigate('catalog', context={'catalog': self})
+        tb_select("Remove Item from the VMDB", invokes_alert=True)
+        sel.handle_alert()
+        flash.assert_success_message('Catalog "{}": Delete successful'.format(self.description))
